+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 Adding notes to what is being changed from time to time
 
+
+## [0.6.0] 10.04.2023
+### Added
+- `en_core_ro_lg` as default model
+- synonym expansion from Romanian WordNet
+- for each synonym, a cosine similarity filtering is used to compare against the context vector of each word
+- params to enable / disable Vector Search and synonyms
+- VectorSearcher using nsmlib to find the most similar n-grams to existing keyword phrases
+- number of characters in the stats
+- text cleaner pre-processor
+
+
 ## [0.5.3] 10.04.2023
 ### Added
 - improved matching to use both lemma and lowercase form of words